Перенести создание модели UIAlertaction в presenter Создать файлы для экрана коллекции
