+++ FakeNFT/Scenes /Catalog/Catalog/CatalogPresenter.swift
@@ -14,7 +14,7 @@ protocol CatalogPresenterProtocol: AnyObject {
     func getNftCollections()
     func sortByName()
     func sortByNftCount()
-    func setSortType() -> UIAlertController
+    func makeSortTypeModel() -> SortTypeModel
 }
 
 final class CatalogPresenter: CatalogPresenterProtocol {
@@ -68,31 +68,14 @@ final class CatalogPresenter: CatalogPresenterProtocol {
         catalogView?.reloadCatalogTableView()
     }
     
-    func setSortType() -> UIAlertController {
-        
-        let actionSheet = UIAlertController(
+    func makeSortTypeModel() -> SortTypeModel {
+        let sortTypeModel = SortTypeModel(
             title: "Сортировка",
-            message: nil,
-            preferredStyle: .actionSheet
+            byName: "По названию",
+            byNftCount: "По количеству NFT",
+            close: "Закрыть"
         )
-        
-        actionSheet.addAction(UIAlertAction(
-            title: "По названию",
-            style: .default) { _ in self.sortByName()
-            })
-        actionSheet.addAction(UIAlertAction(
-            title: "По количеству NFT",
-            style: .default) { _ in
-                self.sortByNftCount()
-            })
-        
-        actionSheet.addAction(
-            UIAlertAction(
-                title: "Закрыть",
-                style: .cancel)
-        )
-        
-        return actionSheet
+        return sortTypeModel
     }
     
     // MARK: - Private Methods
