+++ FakeNFT/Scenes /Catalog/CatalogModels/SortType.swift
@@ -1,0 +1,15 @@
+//
+//  SortTypeModel.swift
+//  FakeNFT
+//
+//  Created by Gleb on 15.08.2024.
+//
+
+import Foundation
+
+struct SortTypeModel {
+    let title: String
+    let byName: String
+    let byNftCount: String
+    var close: String
+}
