+++ FakeNFT/Scenes /Catalog/Collection/CollectionPresenter.swift
@@ -1,0 +1,8 @@
+//
+//  CollectionPresenter.swift
+//  FakeNFT
+//
+//  Created by Gleb on 15.08.2024.
+//
+
+import Foundation
