+++ FakeNFT/Scenes /Catalog/Collection/CollectionViewCell.swift
@@ -1,0 +1,8 @@
+//
+//  CollectionViewCell.swift
+//  FakeNFT
+//
+//  Created by Gleb on 15.08.2024.
+//
+
+import Foundation
