+++ FakeNFT/Scenes /Catalog/Collection/CollectionViewController.swift
@@ -1,0 +1,8 @@
+//
+//  CollectionViewController.swift
+//  FakeNFT
+//
+//  Created by Gleb on 15.08.2024.
+//
+
+import Foundation
